requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
flask-socketio
eventlet
requests
Pillow
opencv-python
torch
transformers
gunicorn
pywebpush
cryptography
python-dotenv
google-generativeai
PyJWT
bcrypt
psycopg2-binary
